qcacld-2.0: Pass tx_pkt_fail_cnt_threshold to FW-1
Add data structures to save tx_pkt_fail_cnt_threshold and changes
to pass the same to FW.

CRs-Fixed: 1041575

+++ CORE/SERVICES/WMA/wma.c
@@ -30216,7 +30216,7 @@ static VOS_STATUS wma_update_tx_fail_cnt_th(tp_wma_handle wma,
 		struct sme_tx_fail_cnt_threshold *tx_fail_cnt_th)
 {
 	u_int8_t vdev_id;
-	u_int16_t tx_fail_disconn_th;
+	u_int32_t tx_fail_disconn_th;
 	int ret = -EIO;
 
 	if (!wma || !wma->wmi_handle) {
